feat: Navigation Stack 구조 분리
