+++ src/navigation/types.ts
@@ -1,0 +1,25 @@
+export type HomeStackParamList = {
+  Home: undefined;
+  ReportList: undefined;
+  ReportArticleDetail: undefined;
+};
+
+export type LoggedOutStackParamList = {
+  Onboarding: undefined;
+  DisasterNotiSettings: undefined;
+  SignIn: undefined;
+  SignUp: undefined;
+  SetName: undefined;
+  SelectLocation: undefined;
+  CompleteLogin: undefined;
+};
+
+export type RootStackParamList = {
+  MainTabNavigator: undefined;
+};
+
+export type RootTabParamList = {
+  Main: undefined;
+  Report: undefined;
+  Solution: undefined;
+};
