+++ src/pages/Main.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import HomeStack from '../navigation/HomeStack';
+
+export default function Main() {
+  return <HomeStack />;
+}
